basic_node and basic_node_iterator, changes to Node
Created a basic_node struct and basic_node_iterator to allow the basic
iterator operators pre++, post++, !=, and == to NOT be templated. This
should save significant space in compiled code as they will not have to
be included for each type.

Changed next member of Node to be next_node to allow interface method
for accessing internal data to be called next.

changed Node get functions to next and data.
They were not resolving b/c the compiler could not determine which to
use when implementing them as overloaded functions with the same input
arguments.

updates to main test method